src/serveur/social/pyproject.toml
[project]
name = "demperm-social"
version = "1.0.0"
description = "Secure social network backend with JWT authentication"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django==5.2.7",
    "djangorestframework>=3.16.1",
    "pyjwt>=2.8.0",
    "bleach>=6.1.0",
    "redis>=5.0.0",
    "django-redis>=5.4.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
]
